Added No TNT Duping
It's not my place to edit the rules, so ignore, but i think adding that to rule 5 would be beneficial, since this server is affiliated with Mumbo, who does use that exploit, so new players might not understand it is also against the rules. No actual changes to the rules, just a detail added.